Allow apertures not == 1 mm^2, account for photons not parallel ot the optical axis
closes #11
