+++ marxs/optics/tests/test_optics.py
@@ -11,16 +11,15 @@ import marxs.optics
 import marxs.optics.base
 import marxs.optics.aperture
 import marxs.source
-import marxs.source.source
 from marxs.utils import generate_test_photons
 
 @pytest.fixture(autouse=True)
 def photons1000():
     '''Make a list of photons parallel to optical axis'''
-    mysource = marxs.source.source.PointSource((30., 30.), energy=1., flux=300.)
+    mysource = marxs.source.PointSource((30., 30.), energy=1., flux=300.)
     np.random.seed(0)
     p = mysource.generate_photons(1000)
-    mypointing = marxs.source.source.FixedPointing(coords=(30., 30.))
+    mypointing = marxs.source.FixedPointing(coords=(30., 30.))
     return mypointing.process_photons(p)
