+++ marxs/source/__init__.py
@@ -1,5 +1,6 @@
-from .source import (Source, PointSource, FixedPointing, JitterPointing,
+from .source import (Source, PointSource,
                      SymbolFSource,
                      poisson_process,
                      )
+from .pointing import FixedPointing, JitterPointing
 from .labSource import LabPointSource, FarLabPointSource, LabPointSourceCone
